(#4701) - extract out browser-only code
