+++ lib/adapters-browser.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = {
+  idb: require('./adapters/idb'),
+  websql: require('./adapters/websql')
+};
