+++ lib/adapters.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = {
+  leveldb: require('./adapters/leveldb')
+};
