+++ lib/deps/binary/binaryStringToBlobOrBuffer-browser.js
@@ -1,0 +1,8 @@
+'use strict';
+
+var createBlob = require('./blob');
+var binaryStringToArrayBuffer = require('./binaryStringToArrayBuffer');
+
+module.exports = function (binString, type) {
+  return createBlob([binaryStringToArrayBuffer(binString)], {type: type});
+};
