+++ lib/deps/binary/buffer.js
@@ -1,2 +1,2 @@
-//this soley exists so we can exclude it in browserify
+//this solely exists so we can exclude it in browserify
 module.exports = Buffer;
