+++ lib/index.js
@@ -14,11 +14,10 @@ var httpAdapter = require('./adapters/http');
 PouchDB.adapter('http', httpAdapter);
 PouchDB.adapter('https', httpAdapter);
 
-PouchDB.adapter('idb', require('./adapters/idb'), true);
-PouchDB.adapter('websql', require('./adapters/websql'), true);
 PouchDB.plugin(require('./mapreduce'));
 
-if (!process.browser) {
-  var ldbAdapter = require('./adapters/leveldb');
-  PouchDB.adapter('leveldb', ldbAdapter, true);
-}
+var adapters = require('./adapters');
+
+Object.keys(adapters).forEach(function (adapterName) {
+  PouchDB.adapter(adapterName, adapters[adapterName], true);
+});
