+++ lib/mapreduce/md5.js
@@ -1,0 +1,7 @@
+'use strict';
+
+var crypto = require('crypto');
+
+module.exports = function (string) {
+  return crypto.createHash('md5').update(string).digest('hex');
+};
