+++ lib/mapreduce/utils.js
@@ -93,14 +93,4 @@ exports.uniq = function (arr) {
   return output;
 };
 
-var crypto = require('crypto');
-var Md5 = require('spark-md5');
-
-exports.MD5 = function (string) {
-  /* istanbul ignore else */
-  if (!process.browser) {
-    return crypto.createHash('md5').update(string).digest('hex');
-  } else {
-    return Md5.hash(string);
-  }
-};
+exports.MD5 = require('./md5');
