+++ package.json
@@ -114,6 +114,11 @@
     "./lib/adapters/preferredAdapters.js": "./lib/adapters/preferredAdapters-browser.js",
     "./lib/deps/binary/buffer.js": "./lib/deps/binary/buffer-browser.js",
     "./lib/deps/migrate.js": "./lib/deps/migrate-browser.js",
+    "./lib/adapters.js": "./lib/adapters-browser.js",
+    "./lib/deps/binary/base64StringToBlobOrBuffer.js": "./lib/deps/binary/base64StringToBlobOrBuffer-browser.js",
+    "./lib/deps/binary/binaryStringToBlobOrBuffer.js": "./lib/deps/binary/binaryStringToBlobOrBuffer-browser.js",
+    "./lib/mapreduce/md5.js": "./lib/mapreduce/md5-browser.js",
+    "./lib/deps/md5.js": "./lib/deps/md5-browser.js",
     "fs": false,
     "leveldown": false
   }
